requirements.txt
Django==5.0.1
psycopg2-binary>=2.9.10
django-environ==0.11.2
django-modeltranslation>=0.19.0
django-ckeditor==6.7.0
django-storages==1.14.2
boto3==1.34.34
sendgrid==6.11.0
django-ratelimit==4.1.0
django-honeypot==0.9.0
django-import-export==3.3.4
Pillow>=10.4.0
python-decouple==3.8
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0

